ai-career-planner/backend/requirements.txt
fastapi
uvicorn
pydantic-ai
python-dotenv
httpx
logfire
pydantic
